Issue of Longclick listener working on websearch is resolved.
